+++ src/cc3002.tarea1/PokemonTypes/WaterPokemon.java
@@ -3,7 +3,7 @@ import cc3002.tarea1.Pokemon;
 import cc3002.tarea1.ISkill;
 import java.util.ArrayList;
 public class WaterPokemon extends Pokemon{
-    public WaterPokemon(String name, int id, double healthPoints, ArrayList<ISkill> skills){
+    public WaterPokemon(String name, int id, int healthPoints, ArrayList<ISkill> skills){
         super(name, id, healthPoints, skills);
     }
     public void attack(Pokemon A){
